replace grabbable with pointable
