+++ examples/laser-physics/index.html
@@ -29,7 +29,7 @@
                   static-body="shape: sphere; sphereRadius: 0.02;">
         </a-mixin>
         <a-mixin id="cube" geometry="primitive: box; width: 0.5; height: 0.5; depth: 0.5"
-                 hoverable pointable drag-droppable
+                 hoverable grabbable drag-droppable
                  shadow dynamic-body
                  event-set="_event: drag-drop; geometry.primitive: sphere; geometry.radius: 0.25"></a-mixin>
         <a-mixin id="cube-hovered" material="opacity: 0.7; transparent: true">
