+++ examples/laser/index.html
@@ -27,7 +27,7 @@
                  raycaster="objects: .cube">
         </a-mixin>
         <a-mixin id="cube" geometry="primitive: box; width: 0.5; height: 0.5; depth: 0.5"
-                 hoverable pointable drag-droppable
+                 hoverable grabbable drag-droppable
                  shadow
                  event-set="_event: drag-drop; geometry.primitive: sphere; geometry.radius: 0.25"></a-mixin>
         <a-mixin id="cube-hovered" material="opacity: 0.7; transparent: true">
