+++ examples/mouse/index.html
@@ -29,7 +29,7 @@
              touch-to-click-converter physics="gravity: 0">
       <a-assets>
         <a-mixin id="cube" geometry="primitive: box; width: 0.5; height: 0.5; depth: 0.5"
-                 hoverable pointable drag-droppable
+                 hoverable grabbable drag-droppable
                  shadow
                  dynamic-body
                  event-set="_event: drag-drop; geometry.primitive: sphere; geometry.radius: 0.25"></a-mixin>
