+++ machinima_tests/scenes/hands-laser.html
@@ -15,7 +15,7 @@
                               dragDropEndButtons: mouseup;"
                  sphere-collider="objects: .cube, a-link"></a-mixin>
         <a-mixin id="cube" geometry="primitive: box; width: 0.5; height: 0.5; depth: 0.5"
-                 hoverable pointable
+                 hoverable grabbable
                  event-set="_event: drag-drop; geometry.primitive: sphere; geometry.radius: 0.25"></a-mixin>
         <a-mixin id="cube-hovered" material="opacity: 0.7; transparent: true">
         <a-mixin id="cube-grabbed" material="opacity: 0.7; transparent: true">
