+++ machinima_tests/super_hands/super-hands-machinima.test.js
@@ -221,7 +221,7 @@ suite('Locomotion', function () {
     this.scene.addEventListener('replayingstopped', e => {
       let z = document.querySelector('[camera]')
         .object3DMap.camera.getWorldPosition().z;
-      assert.isBelow(z, 0.25, 'camera ending z position');
+      assert.isBelow(z, 0.5, 'camera ending z position');
       done();
     }, { once: true }); // once flag because this event emitted multiple times
   });
@@ -292,7 +292,7 @@ suite('camera userHeight', function () {
   });
 });
 
-suite('laser-controls pointable', function () {
+suite('laser-controls grabbable', function () {
   this.timeout(0); // disable Mocha timeout within tests
   setup(function (done) {
     /* inject the scene html into the testing docoument */
